export new hooks and components, organize indexes
