+++ packages/react/src/components/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useIdetik } from "./useIdetik";
+export { useOmeZarrViewer } from "./useOmeZarrImageViewer";
