+++ packages/react/src/components/viewers/OmeZarrImageViewer/index.tsx
@@ -1,1 +1,2 @@
 export { OmeZarrImageViewer } from "./OmeZarrImageViewer";
+export { ChannelControlsList } from "./components/ChannelControlsList";
