+++ packages/react/src/index.ts
@@ -1,7 +1,18 @@
 // import css so it gets packaged with the library
 import "./input.css";
 import { Renderer } from "./components/viewers/OmeZarrImageViewer/components/Renderer";
-import { OmeZarrImageViewer } from "./components/viewers/OmeZarrImageViewer";
-import { useOmeZarrViewer } from "./components/viewers/OmeZarrImageViewer/useOmeZarrImageViewer";
+import {
+  OmeZarrImageViewer,
+  ChannelControlsList,
+} from "./components/viewers/OmeZarrImageViewer";
+import { useIdetik, useOmeZarrViewer } from "./components/hooks";
+import { IdetikProvider } from "./components/providers";
 
-export { Renderer, OmeZarrImageViewer, useOmeZarrViewer };
+export {
+  Renderer,
+  OmeZarrImageViewer,
+  ChannelControlsList,
+  useIdetik,
+  useOmeZarrViewer,
+  IdetikProvider,
+};
